refactor(composer_rpc): move build_and_run_l1_postbatch_trace to delivery.rs
Move the L1 postBatch simulation function (318 LOC) from l1_to_l2.rs
to the shared delivery.rs module. Used by the L1→L2 post-discovery
enrichment for iterative retrace and post-convergence bottom-up
enrichment.

l1_to_l2.rs: 4156 → 3828 (-328 lines).
delivery.rs: 804 → 1132 (+328 lines).

Progress: l1_to_l2 4345→3828 (-517), l2_to_l1 5459→4170 (-1289).

616/616 tests, clippy clean.
