+++ crates/based-rollup/src/composer_rpc/delivery.rs
@@ -9,7 +9,8 @@ use alloy_primitives::{Address, U256};
 use alloy_sol_types::SolCall;
 use serde_json::Value;
 
-use super::model::{L1ProxyLookup, ReturnEdge};
+use super::common::{get_l1_block_context, get_rollup_state_root, get_verification_key};
+use super::model::{DiscoveredCall, L1ProxyLookup, ReturnEdge};
 use crate::cross_chain::{self, ScopePath, filter_new_by_count};
 
 /// Maximum number of iterative discovery rounds in `simulate_l1_delivery`.
@@ -801,4 +802,331 @@ pub(crate) async fn extract_l1_to_l2_return_calls(
             }
         })
         .collect()
+}
+
+/// Build L1 execution entries from detected calls and run a `debug_traceCallMany`
+/// bundle on L1: `[postBatch(entries), userTx]`. Returns the user tx trace (bundle[0][1])
+/// and the full JSON-RPC response, or `None` on failure.
+///
+/// This encapsulates the entry-building, proof-signing, and traceCallMany execution
+/// that is shared between the iterative discovery loop and the post-convergence
+/// enrichment pass. Callers provide `label` for log attribution.
+#[allow(clippy::too_many_arguments)]
+pub(crate) async fn build_and_run_l1_postbatch_trace(
+    client: &reqwest::Client,
+    l1_rpc_url: &str,
+    rollups_address: Address,
+    rollup_id: u64,
+    builder_key: &alloy_signer_local::PrivateKeySigner,
+    detected_calls: &[DiscoveredCall],
+    user_from: &str,
+    user_to: &str,
+    user_data: &str,
+    user_value: &str,
+    label: &str,
+) -> Option<(Value, Value)> {
+    // Build L1DetectedCall entries from known calls
+    let l1_detected: Vec<crate::table_builder::L1DetectedCall> = detected_calls
+        .iter()
+        .map(|c| crate::table_builder::L1DetectedCall {
+            destination: c.destination,
+            data: c.calldata.clone(),
+            value: c.value,
+            source_address: c.source_address,
+            l2_return_data: c.delivery_return_data.clone(),
+            call_success: !c.delivery_failed,
+            parent_call_index: c.parent_call_index,
+            target_rollup_id: if c.parent_call_index.is_child() && c.target_rollup_id == 0 {
+                Some(0)
+            } else {
+                None
+            },
+            scope: if c.trace_depth <= 1 {
+                ScopePath::root()
+            } else {
+                ScopePath::from_parts(vec![U256::ZERO; c.trace_depth])
+            },
+            discovery_iteration: c.discovery_iteration,
+            l1_trace_depth: c.trace_depth,
+            in_reverted_frame: c.in_reverted_frame,
+        })
+        .collect();
+
+    let analyzed = crate::table_builder::analyze_continuation_calls(&l1_detected, rollup_id);
+
+    // Log the call tree
+    tracing::info!(
+        target: "based_rollup::l1_proxy",
+        "({label}) call tree: {} calls, {} analyzed", detected_calls.len(), analyzed.len()
+    );
+    for (i, c) in detected_calls.iter().enumerate() {
+        let sel = if c.calldata.len() >= 4 {
+            format!("0x{}", hex::encode(&c.calldata[..4]))
+        } else {
+            "0x".to_string()
+        };
+        tracing::info!(
+            target: "based_rollup::l1_proxy",
+            "  ({label}) CALL[{}]: dest={} src={} sel={} ret_len={} parent={:?}",
+            i, c.destination, c.source_address, sel, c.delivery_return_data.len(), c.parent_call_index
+        );
+    }
+
+    for (i, a) in analyzed.iter().enumerate() {
+        tracing::info!(
+            target: "based_rollup::l1_proxy",
+            label, i,
+            direction = ?a.direction,
+            is_continuation = a.is_continuation,
+            parent = ?a.parent_call_index,
+            depth = a.depth,
+            dest = %a.call_action.destination,
+            value = %a.call_action.value,
+            scope_len = a.scope.len(),
+            delivery_data_len = a.delivery_return_data.len(),
+            l2_return_data_len = a.l2_return_data.len(),
+            "analyzed call"
+        );
+    }
+
+    let entries = if analyzed.is_empty() {
+        // Fallback: build simple CALL+RESULT pairs, then convert to L1 format.
+        let l2_pairs: Vec<_> = l1_detected
+            .iter()
+            .flat_map(|c| {
+                let (call_entry, result_entry) = cross_chain::build_cross_chain_call_entries(
+                    cross_chain::RollupId::new(alloy_primitives::U256::from(rollup_id)),
+                    c.destination,
+                    c.data.clone(),
+                    c.value,
+                    c.source_address,
+                    cross_chain::RollupId::MAINNET,
+                    c.call_success,
+                    c.l2_return_data.clone(),
+                );
+                vec![call_entry, result_entry]
+            })
+            .collect();
+        cross_chain::convert_pairs_to_l1_entries(&l2_pairs)
+    } else {
+        let cont = crate::table_builder::build_continuation_entries(
+            &analyzed,
+            cross_chain::RollupId::new(alloy_primitives::U256::from(rollup_id)),
+        );
+        tracing::info!(
+            target: "based_rollup::l1_proxy",
+            l2_entries = cont.l2_entries.len(),
+            l1_entries = cont.l1_entries.len(),
+            "({label}) built continuation entries"
+        );
+        cont.l1_entries
+    };
+
+    // Log entries with state delta details
+    for (idx, e) in entries.iter().enumerate() {
+        let delta_info = if let Some(d) = e.state_deltas.first() {
+            format!(
+                "rollup={} current={} new={} ether_delta={}",
+                d.rollup_id, d.current_state, d.new_state, d.ether_delta
+            )
+        } else {
+            "no deltas".to_string()
+        };
+        tracing::info!(
+            target: "based_rollup::l1_proxy",
+            label, idx,
+            action_hash = %e.action_hash,
+            next_type = ?e.next_action.action_type,
+            next_dest = %e.next_action.destination,
+            next_scope_len = e.next_action.scope.len(),
+            next_data_len = e.next_action.data.len(),
+            next_value = %e.next_action.value,
+            delta = %delta_info,
+            "L1 entry for postBatch simulation"
+        );
+    }
+
+    // Fix placeholder state deltas for simulation entries.
+    // build_continuation_entries produces entries with placeholder
+    // currentState=0x0 / newState=0x0. _findAndApplyExecution checks
+    // rollups[rollupId].stateRoot == delta.currentState — placeholders
+    // won't match. Query the real on-chain stateRoot and set identity
+    // deltas (current=real, new=real) while preserving ether_delta.
+    let mut entries = entries;
+    let on_chain_state_root = get_rollup_state_root(client, l1_rpc_url, rollups_address, rollup_id)
+        .await
+        .unwrap_or(alloy_primitives::B256::ZERO);
+    for e in &mut entries {
+        for d in &mut e.state_deltas {
+            d.current_state = on_chain_state_root;
+            d.new_state = on_chain_state_root;
+            // ether_delta is preserved from build_continuation_entries
+        }
+    }
+
+    if entries.is_empty() {
+        return None;
+    }
+
+    // Get L1 block context for proof signing.
+    let block_ctx = get_l1_block_context(client, l1_rpc_url).await;
+    let (block_number, block_hash, _parent_hash) = match block_ctx {
+        Ok(ctx) => ctx,
+        Err(e) => {
+            tracing::warn!(target: "based_rollup::l1_proxy", %e, "({label}) failed to get L1 block context");
+            return None;
+        }
+    };
+
+    // Get verification key from Rollups contract
+    let vk = match get_verification_key(client, l1_rpc_url, rollups_address, rollup_id).await {
+        Ok(v) => v,
+        Err(e) => {
+            tracing::warn!(target: "based_rollup::l1_proxy", %e, "({label}) failed to get verification key");
+            return None;
+        }
+    };
+
+    // Sign ECDSA proof for postBatch in traceCallMany context.
+    let trace_parent_hash = block_hash;
+    // `UNIX_EPOCH` is by definition <= now on any system with a sane clock; we
+    // still handle the rare SystemTime-before-epoch case defensively so there
+    // is no `.unwrap()` in production code.
+    let trace_block_timestamp = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .map(|d| d.as_secs())
+        .unwrap_or(0);
+
+    let call_data_bytes = alloy_primitives::Bytes::new();
+    let entry_hashes = cross_chain::compute_entry_hashes(&entries, vk);
+    let public_inputs_hash = cross_chain::compute_public_inputs_hash(
+        &entry_hashes,
+        &call_data_bytes,
+        trace_parent_hash,
+        trace_block_timestamp,
+    );
+
+    use alloy_signer::SignerSync;
+    let sig = match builder_key.sign_hash_sync(&public_inputs_hash) {
+        Ok(s) => s,
+        Err(e) => {
+            tracing::warn!(target: "based_rollup::l1_proxy", %e, "({label}) failed to sign proof");
+            return None;
+        }
+    };
+    let sig_bytes = sig.as_bytes();
+    let mut proof_bytes = sig_bytes.to_vec();
+    if proof_bytes.len() == 65 && proof_bytes[64] < 27 {
+        proof_bytes[64] += 27;
+    }
+    let proof = alloy_primitives::Bytes::from(proof_bytes);
+
+    // Encode postBatch calldata
+    let post_batch_calldata =
+        cross_chain::encode_post_batch_calldata(&entries, call_data_bytes, proof);
+
+    // Build traceCallMany request: [postBatch, userTx] in a single bundle
+    let builder_addr = format!("{}", builder_key.address());
+    let rollups_hex = format!("{rollups_address}");
+    let post_batch_data = format!("0x{}", hex::encode(post_batch_calldata.as_ref()));
+
+    let next_block = format!("{:#x}", block_number + 1);
+    let trace_req = serde_json::json!({
+        "jsonrpc": "2.0",
+        "method": "debug_traceCallMany",
+        "params": [
+            [
+                {
+                    "transactions": [
+                        {
+                            "from": builder_addr,
+                            "to": rollups_hex,
+                            "data": post_batch_data,
+                            "gas": "0x1c9c380"
+                        },
+                        {
+                            "from": user_from,
+                            "to": user_to,
+                            "data": user_data,
+                            "value": user_value,
+                            "gas": "0x2faf080"
+                        }
+                    ],
+                    "blockOverride": {
+                        "number": next_block,
+                        "time": format!("{:#x}", trace_block_timestamp)
+                    }
+                }
+            ],
+            null,
+            { "tracer": "callTracer" }
+        ],
+        "id": 3
+    });
+
+    let resp = match client.post(l1_rpc_url).json(&trace_req).send().await {
+        Ok(r) => match r.json::<Value>().await {
+            Ok(v) => v,
+            Err(e) => {
+                tracing::warn!(target: "based_rollup::l1_proxy", %e, "({label}) traceCallMany response parse failed");
+                return None;
+            }
+        },
+        Err(e) => {
+            tracing::warn!(target: "based_rollup::l1_proxy", %e, "({label}) traceCallMany request failed");
+            return None;
+        }
+    };
+
+    // Extract traces from result.
+    let bundle_traces = match resp
+        .get("result")
+        .and_then(|r| r.get(0))
+        .and_then(|b| b.as_array())
+    {
+        Some(arr) if arr.len() >= 2 => arr,
+        _ => {
+            if let Some(error) = resp.get("error") {
+                tracing::warn!(
+                    target: "based_rollup::l1_proxy",
+                    ?error,
+                    "({label}) traceCallMany returned error"
+                );
+            } else {
+                tracing::warn!(
+                    target: "based_rollup::l1_proxy",
+                    "({label}) traceCallMany returned unexpected structure"
+                );
+            }
+            return None;
+        }
+    };
+
+    // Check if postBatch succeeded
+    let tx1_trace = &bundle_traces[0];
+    if tx1_trace.get("error").is_some() || tx1_trace.get("revertReason").is_some() {
+        let error_msg = tx1_trace
+            .get("error")
+            .and_then(|v| v.as_str())
+            .unwrap_or("unknown");
+        let revert_reason = tx1_trace
+            .get("revertReason")
+            .and_then(|v| v.as_str())
+            .unwrap_or("none");
+        let output = tx1_trace
+            .get("output")
+            .and_then(|v| v.as_str())
+            .unwrap_or("none");
+        tracing::warn!(
+            target: "based_rollup::l1_proxy",
+            error = error_msg,
+            revert_reason,
+            output,
+            "({label}) postBatch reverted in traceCallMany — entries may be invalid"
+        );
+        // Still return the user tx trace — caller may need partial results
+    }
+
+    let user_trace = bundle_traces[1].clone();
+    Some((user_trace, resp))
 }
